(PC-6368) dissociate Landing page from search results page
