+++ src/features/search/components/index.tsx
@@ -6,4 +6,6 @@ export * from './PriceSlider'
 export * from './RadiusSlider'
 export * from './SearchBox'
 export * from './SearchHeader'
+export * from './SearchLandingPage'
+export * from './SearchResults'
 export * from './SectionWithSwitch'
